Reviews complete

+++ server/controllers/reviewController.js
@@ -1,6 +1,8 @@
 const reviewSchema = require("../models/reviewSchema")
+const productSchema = require("../models/productSchema")
 const resHandler = require("../utils/resHandler")
 const { ObjectId } = require('mongodb');
+const updateProductReviewStats = require('../services/updateProductReviewStats');
 
 // ============ create review 
 const createReview = async (req, res) => {
@@ -24,7 +26,10 @@ const createReview = async (req, res) => {
             rating,
             comment
         })
-        newReview.save()
+        await newReview.save()
+
+        // update product review stats
+        await updateProductReviewStats(productId)
 
         // ----------- Success 
         resHandler.success(res, 201, "Product review added")
@@ -43,12 +48,18 @@ const editReview = async (req, res) => {
         if (!rating || rating < 1 || rating > 5) return resHandler.error(res, 400, 'rating is required and must be at between 1-5')
         if (!comment) return resHandler.error(res, 400, 'comment cannot be empty')
 
+        // ensure review exists and get product id
+        const existing = await reviewSchema.findById(reviewId)
+        if (!existing) return resHandler.error(res, 404, 'review not found')
+
         // ------------- Save to DB 
         await reviewSchema.findByIdAndUpdate(reviewId, {
             rating,
             comment
         }, { new: true })
 
+        // update product review stats
+        await updateProductReviewStats(existing.product)
 
         // ----------- Success 
         resHandler.success(res, 200, "Review edited successfully")
@@ -66,8 +77,15 @@ const deleteReview = async (req, res) => {
         if (!reviewId) return resHandler.error(res, 400, 'review id is required')
         if (!ObjectId.isValid(reviewId)) return resHandler.error(res, 400, 'invalid object id')
 
-        // ------------- Save to DB 
+        // get review to obtain product id
+        const existing = await reviewSchema.findById(reviewId)
+        if (!existing) return resHandler.error(res, 404, 'review not found')
+
+        // ------------- Delete from DB 
         await reviewSchema.findByIdAndDelete(reviewId)
+
+        // update product review stats
+        await updateProductReviewStats(existing.product)
 
         // ----------- Success 
         resHandler.success(res, 200, "Review deleted successfully")
@@ -95,6 +113,5 @@ const getByUser = async (req, res) => {
         resHandler.error(res, 500, "Internal server error")
     }
 }
-
 
 module.exports = { createReview, editReview, deleteReview, getByUser }
